requirements.txt
openai>=1.3.0
python-dotenv
pydantic
